fix: fix cards tests
