+++ src/components/Card/Card.test.tsx
@@ -1,6 +1,7 @@
-import { act, render, screen } from '@testing-library/react';
+import { act, screen } from '@testing-library/react';
 import { mockData } from '../../test/mockData';
 import Card from './Card';
+import { renderWithProviders } from '../../utils/test-utils';
 
 describe('Card', () => {
   beforeEach(() => {
@@ -10,7 +11,7 @@ describe('Card', () => {
   it('should render the relevant card data', () => {
     const { title, title_english, synopsis, duration, score } = mockData[0];
     act(() => {
-      render(<Card data={mockData[0]} />);
+      renderWithProviders(<Card data={mockData[0]} />);
     });
     expect(screen.getByText(`${score}`)).toBeInTheDocument();
     expect(screen.getByText(`${duration}`)).toBeInTheDocument();
